Place messaging infrastructure inside a VPC
This commit refactors the terraform deployment to create a VPC into which we create a new subnet, internet gateway and launch our
rabbitmq 2 servers into. This commit explicitly allows internal VPC traffic to the rabbitmq servers including clustering
and amqp connections. We may choose to lock this down further.

We currently still allow ssh connections to the rabbitmq instances for provisioning but we will need to remove this later, moving the
provisioning to work via a bastion box.