Create VPC, submitter ec2 instances and place rabbitmq behind ELB
This commit is a major refactor of the terraform setup.

* It creates a new Elastic beanstalk instance for use by survey runner using the experimental feature branch of Terraform.
* It adds an elastic load balancer in front of two existing rabbitmq ec2 instances.
* Creates two ec2 instances for use by the eq-submitter

It wraps all of the above in a single VPC, but currently makes some of those components available to the public web
so as to allow ssh connections for provisioning. This will be tightened in upcoming stories / PR's.

Filenames were also renamed to create a consistent naming convention and an outputs file to give a clear view of what
a run should output.