Inject environment variables into ElasticBeanstalk and EC2
For both the survey runner and the submitter, this commit creates environment variables for connecting to the Rabbitmq servers.
We also inject some environment settings around what type of environment that apps are operating in and centralise the setting of
some variables (namely the queue name and Rabbitmq url).

We also update the terraform.tfvars.example to allow for setting the submission bucket creds.