Enable TLS support for Elasticbeanstalk
This commit places the elasticbeanstalk environment behind a TLS certificate.

To enable this in each environment (dev, preprod, prod) you need to set a new
variable `certificate_arn` which is a reference to a pre-loaded certificate
in each env.

You can find this out via the AWS console or asking another team member.