Remove secondary network interfaces and staticly set ips of rabbitmq servers
To enable the VPN routing of connectivity to the rabbitmq servers to
work we need to allocate static ip's to our two rabbitmq servers.

Previously we did this by creating two new interfaces and attaching them to the servers.
This commit simplifies the design by just allocating those IP's to the instances
as they come up.