Remove ELB in front of RabbitMQ servers
This commit removes the ELB in front of the Rabbitmq servers as during network partitioning we saw
messages get stuck on a partitioned cluster. This requires a change in our connection logic from
the survey runner to before being deployed to live, as we no longer will automatically failover in
the event of the prime (main) Rabbitmq server failing.