Fix bug in message queue security group allocation
There was a bug in the message queue security group allocation
causing the ec2 instances to always get destroyed and re-created
on each terraform run. Previously we were using the security
group id's with the security_groups property of aws_instance.
Changing this to vpc_security_group_ids fixes the issue.