Add scaling based on CPU utilization and reduce healthcheck timings
Scaling by CPU utilization has proved the most effective metric in load
testing - the application is CPU intensive and usage goes up before the
latency of requests increase. It has been set to scale at 40% as under
increasing load this brought instances into service in time to ensure
latency stayed low.

The healthchecks timings and number we do has been reduced as these
have to be completed before an instance comes into service. Two
healthchecks 5 seconds part is more than enough to determine if an
instance should be in service or not.