Add ansible requirements file
The previous approach of installing the alexey-medvedchikov role with a
fixed git commit didn't work as expected. The role was installed, but as
a role named ansible-rabbitmq, which then wasn't detected by Ansible.
Installing the role via a requirements YAML gives the chance to override
the role name on installation, and also means the roles and versions are
in configuration rather than just the README.

ansible.cfg has also been symlinked from the queue module to the
eq-terraform directory so it is read when running Terraform from that
directory