Removed control_path from Ansible config
It was causing a '[string] too long for unix domain socket'
error. After some googling it turns out control path was often used as
a workaround for this error but now a proper fix has gone in and
having it can actually cause the error. The recommendation is to
remove it which worked for me. See the end of https://github.com/ansible/ansible/issues/11536.