Fixed ansible run issue when hostname has trailing dot (#95)
* Fixed ansible run issue when hostname has trailing dot

* Enable Detailed Monitoring

* Remove duplicated alarm and use count

* Treat missing data as breaching

* Tidy up Alerting

* Increase cluster status alarm to 2 evaluation_periods

* Removed control_path from Ansible config

It was causing a '[string] too long for unix domain socket'
error. After some googling it turns out control path was often used as
a workaround for this error but now a proper fix has gone in and
having it can actually cause the error. The recommendation is to
remove it which worked for me. See the end of https://github.com/ansible/ansible/issues/11536.

* Opened Port for SDX

* Disabled Ansible host key checking

Added the environment variable to the front of the playbook command.

* Remove ansible.cfg as it did nothing